[FTR][SLOT-308]: Implement code review feedback. Modify plansResponse to include totalFuturePrices and delete current util. Fix accesor for currentPrice sort
